MappedTransferOperatorsTest moved to OperatorsModel MappedTransferOperatorAndFunctionOfEnumTest and OperatorAndFunctionOfEnumTest moved to OperatorsModel
