--- judo-runtime-core-jsl-itest/models/MappedTransferOperatorAndFunctionOfEnumTestModel/src/test/resources/MappedTransferOperatorAndFunctionOfEnumTest.jsl
@@ -1,36 +1,0 @@
-model MappedTransferOperatorAndFunctionOfEnumTest;
-
-type boolean Bool;
-type string String min-size:0 max-size:10;
-
-enum TestLiteral {
-    Aaa01 = 1;
-    Bbb02 = 2;
-    Ccc03 = 9999;
-    ZZ00  = 0;
-}
-
-entity E1 {
-    field TestLiteral f1 = TestLiteral#ZZ00;
-    field TestLiteral f2 = TestLiteral#Aaa01;
-    field Bool   f3 <= self.f1 < self.f2;
-    field Bool   f4 <= self.f1 <= self.f2;
-    field Bool   f5 <= self.f1 > self.f2;
-    field Bool   f6 <= self.f1 >= self.f2;
-    field Bool   f7 <= self.f1 == self.f2;
-    field Bool   f8 <= self.f1 != self.f2;
-    field String f9 <= self.f1!asString();
-}
-
-transfer TE1 (E1 mapped){
-    field TestLiteral f1 <=> mapped.f1;
-    field TestLiteral f2  <=> mapped.f2;
-    field Bool   f3 <= mapped.f3;
-    field Bool   f4 <= mapped.f4;
-    field Bool   f5 <= mapped.f5;
-    field Bool   f6 <= mapped.f6;
-    field Bool   f7 <= mapped.f7;
-    field Bool   f8 <= mapped.f8;
-    field String f9 <= mapped.f9;
-}
-
