--- judo-runtime-core-jsl-itest/models/OperatorAndFunctionOfEnumTestModel/src/test/resources/OperatorAndFunctionOfEnumTest.jsl
@@ -1,23 +1,0 @@
-model OperatorAndFunctionOfEnumTest;
-
-type boolean Bool;
-type string String min-size:0 max-size:10;
-
-enum TestLiteral {
-    Aaa01 = 1;
-    Bbb02 = 2;
-    Ccc03 = 9999;
-    ZZ00  = 0;
-}
-
-entity E1 {
-    field TestLiteral f1 = TestLiteral#ZZ00;
-    field TestLiteral f2 = TestLiteral#Aaa01;
-    field Bool   f3 <= self.f1 < self.f2;
-    field Bool   f4 <= self.f1 <= self.f2;
-    field Bool   f5 <= self.f1 > self.f2;
-    field Bool   f6 <= self.f1 >= self.f2;
-    field Bool   f7 <= self.f1 == self.f2;
-    field Bool   f8 <= self.f1 != self.f2;
-    field String f9 <= self.f1!asString();
-}
