+++ judo-runtime-core-jsl-itest/models/OperatorsModel/src/test/java/hu/blackbelt/judo/runtime/core/jsl/transfer/MappedTransferOperatorsTest.java
@@ -22,18 +22,24 @@ package hu.blackbelt.judo.runtime.core.jsl.transfer;
 
 import com.google.inject.Inject;
 import com.google.inject.Module;
-import hu.blackbelt.judo.psm.generator.sdk.core.test.api.mappedtransferoperators.mappedtransferoperators.defaultoperators.DefaultOperators;
-import hu.blackbelt.judo.psm.generator.sdk.core.test.api.mappedtransferoperators.mappedtransferoperators.defaultoperators.DefaultOperatorsIdentifier;
-import hu.blackbelt.judo.psm.generator.sdk.core.test.api.mappedtransferoperators.mappedtransferoperators.derivedoperators.DerivedOperators;
-import hu.blackbelt.judo.psm.generator.sdk.core.test.api.mappedtransferoperators.mappedtransferoperators.derivedoperators.DerivedOperatorsIdentifier;
-import hu.blackbelt.judo.psm.generator.sdk.core.test.api.mappedtransferoperators.mappedtransferoperators.transferdefaultoperators.TransferDefaultOperators;
-import hu.blackbelt.judo.psm.generator.sdk.core.test.api.mappedtransferoperators.mappedtransferoperators.transferdefaultoperators.TransferDefaultOperatorsDao;
-import hu.blackbelt.judo.psm.generator.sdk.core.test.api.mappedtransferoperators.mappedtransferoperators.transferderivedoperators.TransferDerivedOperators;
-import hu.blackbelt.judo.psm.generator.sdk.core.test.api.mappedtransferoperators.mappedtransferoperators.transferderivedoperators.TransferDerivedOperatorsDao;
-import hu.blackbelt.judo.psm.generator.sdk.core.test.api.mappedtransferoperators.mappedtransferoperators.transferderivedsource.TransferDerivedSource;
-import hu.blackbelt.judo.psm.generator.sdk.core.test.api.mappedtransferoperators.mappedtransferoperators.defaultoperators.DefaultOperatorsDao;
-import hu.blackbelt.judo.psm.generator.sdk.core.test.api.mappedtransferoperators.mappedtransferoperators.derivedoperators.DerivedOperatorsDao;
-import hu.blackbelt.judo.psm.generator.sdk.core.test.guice.MappedTransferOperatorsDaoModules;
+import hu.blackbelt.judo.psm.generator.sdk.core.test.api.operators.operators.defaultoperators.DefaultOperators;
+import hu.blackbelt.judo.psm.generator.sdk.core.test.api.operators.operators.defaultoperators.DefaultOperatorsIdentifier;
+import hu.blackbelt.judo.psm.generator.sdk.core.test.api.operators.operators.derivedoperators.DerivedOperators;
+import hu.blackbelt.judo.psm.generator.sdk.core.test.api.operators.operators.derivedoperators.DerivedOperatorsIdentifier;
+import hu.blackbelt.judo.psm.generator.sdk.core.test.api.operators.operators.e1.E1;
+import hu.blackbelt.judo.psm.generator.sdk.core.test.api.operators.operators.e1.E1Dao;
+import hu.blackbelt.judo.psm.generator.sdk.core.test.api.operators.operators.e1.E1Identifier;
+import hu.blackbelt.judo.psm.generator.sdk.core.test.api.operators.operators.te1.TE1;
+import hu.blackbelt.judo.psm.generator.sdk.core.test.api.operators.operators.te1.TE1Dao;
+import hu.blackbelt.judo.psm.generator.sdk.core.test.api.operators.operators.testliteral.TestLiteral;
+import hu.blackbelt.judo.psm.generator.sdk.core.test.api.operators.operators.transferdefaultoperators.TransferDefaultOperators;
+import hu.blackbelt.judo.psm.generator.sdk.core.test.api.operators.operators.transferdefaultoperators.TransferDefaultOperatorsDao;
+import hu.blackbelt.judo.psm.generator.sdk.core.test.api.operators.operators.transferderivedoperators.TransferDerivedOperators;
+import hu.blackbelt.judo.psm.generator.sdk.core.test.api.operators.operators.transferderivedoperators.TransferDerivedOperatorsDao;
+import hu.blackbelt.judo.psm.generator.sdk.core.test.api.operators.operators.transferderivedsource.TransferDerivedSource;
+import hu.blackbelt.judo.psm.generator.sdk.core.test.api.operators.operators.defaultoperators.DefaultOperatorsDao;
+import hu.blackbelt.judo.psm.generator.sdk.core.test.api.operators.operators.derivedoperators.DerivedOperatorsDao;
+import hu.blackbelt.judo.psm.generator.sdk.core.test.guice.OperatorsDaoModules;
 import hu.blackbelt.judo.requirement.report.annotation.Requirement;
 import hu.blackbelt.judo.requirement.report.annotation.TestCase;
 import hu.blackbelt.judo.runtime.core.jsl.AbstractJslTest;
@@ -43,6 +49,7 @@ import org.junit.jupiter.api.Test;
 import java.util.Optional;
 
 import static org.junit.jupiter.api.Assertions.*;
+import static org.junit.jupiter.api.Assertions.assertFalse;
 
 @Slf4j
 public class MappedTransferOperatorsTest extends AbstractJslTest {
@@ -59,14 +66,18 @@ public class MappedTransferOperatorsTest extends AbstractJslTest {
     @Inject
     TransferDerivedOperatorsDao transferDerivedOperatorsDao;
 
+    @Inject TE1Dao te1Dao;
+
+    @Inject E1Dao e1Dao;
+
     @Override
     public Module getModelDaoModule() {
-        return new MappedTransferOperatorsDaoModules();
+        return new OperatorsDaoModules();
     }
 
     @Override
     public String getModelName() {
-        return "MappedTransferOperators";
+        return "Operators";
     }
 
     /**
@@ -78,7 +89,7 @@ public class MappedTransferOperatorsTest extends AbstractJslTest {
      *
      * @others Implement this test case in the *judo-runtime-core-jsl-itest* module.
      *
-     * @jslModel MappedTransferOperationsTest.jsl
+     * @jslModel OperationsTest.jsl
      *
      * @positiveRequirements REQ-EXPR-021
      *
@@ -261,7 +272,7 @@ public class MappedTransferOperatorsTest extends AbstractJslTest {
      *
      * @others Implement this test case in the *judo-runtime-core-jsl-itest* module.
      *
-     * @jslModel MappedTransferOperationsTest.jsl
+     * @jslModel OperationsTest.jsl
      *
      * @positiveRequirements REQ-EXPR-021
      *
@@ -360,4 +371,156 @@ public class MappedTransferOperatorsTest extends AbstractJslTest {
         assertTrue(operators.getConditionalUndefinedTrue().orElseThrow());
         assertFalse(operators.getConditionalUndefinedFalse().orElseThrow());
     }
+
+    /**
+     * Testing the operators and functions of enum on a transfer object.
+     *
+     * @prerequisites Nothing
+     *
+     * @type Behaviour
+     *
+     * @others Implement this test case in the *judo-runtime-core-jsl-itest* module.
+     *
+     * @scenario
+     *  . Parse (and/or build) the model.
+     *
+     *  . The result of the model parsing (and/or building) is successful.
+     *
+     *  . Create an E1 entity instance (e1) with the default values.
+     *
+     *  . Check the values of the following fields of the new entity instance (e1).
+     *     * f1 == TestLiteral#ZZ00
+     *     * f2 == TestLiteral#Aaa01
+     *     * f3 == true
+     *     * f4 == true
+     *     * f5 == false
+     *     * f6 == false
+     *     * f7 == false
+     *     * f8 == true
+     *     * f9 == "ZZ00"
+     *
+     *  . Change the e1 entity instance f1 field to TestLiteral#Aaa01 and save the e1 entity instance.
+     *
+     *  . Check the values of the following fields of the modified e1 entity instance.
+     *     * f1 == TestLiteral#Aaa01
+     *     * f2 == TestLiteral#Aaa01
+     *     * f3 == false
+     *     * f4 == true
+     *     * f5 == false
+     *     * f6 == true
+     *     * f7 == true
+     *     * f8 == false
+     *     * f9 == "Aaa01"
+     *
+     *  . Change the e1 entity instance's f1 field to TestLiteral#Ccc03 and f2 field to TestLiteral#ZZ00 and save the e1 entity instance.
+     *
+     *  . Check the values of the following fields of the modified e1 entity instance.
+     *     * f1 == TestLiteral#Ccc03
+     *     * f2 == TestLiteral#ZZ00
+     *     * f3 == false
+     *     * f4 == false
+     *     * f5 == true
+     *     * f6 == true
+     *     * f7 == false
+     *     * f8 == true
+     *     * f9 == "Ccc03"
+     *
+     *  . The test is passed if all checks are successful, and there were no exceptions.
+     */
+    @Test
+    @TestCase("EnumOperatorsAndFunctionsOnTransfer")
+    @Requirement(reqs = {
+            "REQ-SYNT-001",
+            "REQ-SYNT-002",
+            "REQ-SYNT-003",
+            "REQ-SYNT-004",
+            "REQ-MDL-001",
+            "REQ-TYPE-002",
+            "REQ-TYPE-004",
+            "REQ-TYPE-006",
+            "REQ-ENT-001",
+            "REQ-ENT-002",
+            "REQ-ENT-008",
+            "REQ-EXPR-020",
+            "REQ-SRV-002"
+    })
+    void testEnumOperatorsAndFunctionsOnTransfer() {
+
+        TE1 te1 = te1Dao.create(TE1.builder().build());
+        assertEquals(TestLiteral.ZZ00, te1.getF1().orElseThrow());
+        assertEquals(TestLiteral.Aaa01, te1.getF2().orElseThrow());
+        assertTrue(te1.getF3().orElseThrow());
+        assertTrue(te1.getF4().orElseThrow());
+        assertFalse(te1.getF5().orElseThrow());
+        assertFalse(te1.getF6().orElseThrow());
+        assertFalse(te1.getF7().orElseThrow());
+        assertTrue(te1.getF8().orElseThrow());
+        assertEquals("ZZ00", te1.getF9().orElseThrow());
+
+        // entity representation
+
+        E1 e1 = e1Dao.getById(te1.identifier().adaptTo(E1Identifier.class)).orElseThrow();
+        assertEquals(TestLiteral.ZZ00, e1.getF1().orElseThrow());
+        assertEquals(TestLiteral.Aaa01, e1.getF2().orElseThrow());
+        assertTrue(e1.getF3().orElseThrow());
+        assertTrue(e1.getF4().orElseThrow());
+        assertFalse(e1.getF5().orElseThrow());
+        assertFalse(e1.getF6().orElseThrow());
+        assertFalse(e1.getF7().orElseThrow());
+        assertTrue(e1.getF8().orElseThrow());
+        assertEquals("ZZ00", e1.getF9().orElseThrow());
+
+
+        te1.setF1(TestLiteral.Aaa01);
+        te1 = te1Dao.update(te1);
+        assertEquals(TestLiteral.Aaa01, te1.getF1().orElseThrow());
+        assertEquals(TestLiteral.Aaa01, te1.getF2().orElseThrow());
+        assertFalse(te1.getF3().orElseThrow());
+        assertTrue(te1.getF4().orElseThrow());
+        assertFalse(te1.getF5().orElseThrow());
+        assertTrue(te1.getF6().orElseThrow());
+        assertTrue(te1.getF7().orElseThrow());
+        assertFalse(te1.getF8().orElseThrow());
+        assertEquals("Aaa01", te1.getF9().orElseThrow());
+
+        // entity rep
+        e1 = e1Dao.getById(te1.identifier().adaptTo(E1Identifier.class)).orElseThrow();
+        assertEquals(TestLiteral.Aaa01, e1.getF1().orElseThrow());
+        assertEquals(TestLiteral.Aaa01, e1.getF2().orElseThrow());
+        assertFalse(e1.getF3().orElseThrow());
+        assertTrue(e1.getF4().orElseThrow());
+        assertFalse(e1.getF5().orElseThrow());
+        assertTrue(e1.getF6().orElseThrow());
+        assertTrue(e1.getF7().orElseThrow());
+        assertFalse(e1.getF8().orElseThrow());
+        assertEquals("Aaa01", e1.getF9().orElseThrow());
+
+
+        te1.setF1(TestLiteral.Ccc03);
+        te1.setF2(TestLiteral.ZZ00);
+        te1 = te1Dao.update(te1);
+        assertEquals(TestLiteral.Ccc03, te1.getF1().orElseThrow());
+        assertEquals(TestLiteral.ZZ00, te1.getF2().orElseThrow());
+        assertFalse(te1.getF3().orElseThrow());
+        assertFalse(te1.getF4().orElseThrow());
+        assertTrue(te1.getF5().orElseThrow());
+        assertTrue(te1.getF6().orElseThrow());
+        assertFalse(te1.getF7().orElseThrow());
+        assertTrue(te1.getF8().orElseThrow());
+        assertEquals("Ccc03", te1.getF9().orElseThrow());
+
+        // entity representation
+        e1 = e1Dao.getById(te1.identifier().adaptTo(E1Identifier.class)).orElseThrow();
+        assertEquals(TestLiteral.Ccc03, e1.getF1().orElseThrow());
+        assertEquals(TestLiteral.ZZ00, e1.getF2().orElseThrow());
+        assertFalse(e1.getF3().orElseThrow());
+        assertFalse(e1.getF4().orElseThrow());
+        assertTrue(e1.getF5().orElseThrow());
+        assertTrue(e1.getF6().orElseThrow());
+        assertFalse(e1.getF7().orElseThrow());
+        assertTrue(e1.getF8().orElseThrow());
+        assertEquals("Ccc03", e1.getF9().orElseThrow());
+
+    }
 }
+
