+++ judo-runtime-core-jsl-itest/models/pom.xml
@@ -32,9 +32,7 @@
         <module>MappedTransferAssociationAssociationModel</module>
         <module>MappedTransferCompositonAggregationModel</module>
         <module>MappedTransferCompositonAssociationModel</module>
-        <module>MappedTransferNavigationTestModel</module>
         <module>MappedTransferOperatorAndFunctionOfEnumTestModel</module>
-        <module>MappedTransferOperatorsModel</module>
         <module>MappedTransferPrimitivesModel</module>
         <module>ModelNameWithJavaKeyword</module>
         <module>NavigationTestModel</module>
